requirements.txt
requests   
gtts         
openai-whisper  
python-dotenv
pydub
librosa
scipy
markdown
bs4
fastapi
sqlalchemy
psycopg2
httpx
aiosmtplib
uvicorn
alembic
pydantic
pydantic[email]
python-multipart
